+++ CHANGELOG.md
@@ -10,6 +10,7 @@ AdminNeo 5.6.0
 - MariaDB: Support inet4 and inet6 types (by @vrana)
 - MariaDB 11.7+: Support vector columns and indexes (by @vrana)
 - PostgreSQL 9.2+: Highlight queries in process list (by @vrana)
+- SQLite: Support STRICT and WITHOUT ROWID in create and alter table (by @vrana)
 
 ### Bugfixes